# perf: logger module loading