### What I did

improve the loading time for `ape.logging` module

Before:

```
In [1]: %time from ape.logging import *
CPU times: user 73.7 ms, sys: 13.6 ms, total: 87.3 ms
Wall time: 86.2 ms
```

After:

```
In [1]: %time from ape.logging import *
CPU times: user 15.4 ms, sys: 2.55 ms, total: 18 ms
Wall time: 16.9 ms
```

### How I did it

1. rm usage of yarl
2. delay import from `rich`

notes:

1. by doing this, we can safely import from `ape.logging` and `ape.exceptions` anywhere w/o that ever being the bottleneck.
2. this will make loading from `ape.exceptions` also faster (combined with the other PR open)
3. this makes `ape --help` ever so slightly  faster on average.
4. this works towards a refactor which will make a lot of the ugliness go away

### How to verify it

<!-- Discuss any methods that should be used to verify the change -->

### Checklist

<!-- All PRs must complete the following checklist before being merged -->

- [ ] All changes are completed
- [ ] New test cases have been added
- [ ] Documentation has been updated